requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.0
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
pycparser==2.22
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2024.2
